Moves GenServer from UserManagement to UsersRepository
The (upcoming) authorization library will need to access the users from
the repository. The way it was before this, it would end up having to
call back to the UserManagement process for the shared state, which is
not allowed (since the second call would be waiting on the first call to
finish).